+++ src/main/java/lotto/view/LottoScreen.java
@@ -4,9 +4,9 @@ import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.LottoFinalResult;
 import lotto.domain.UserMoney;
-import lotto.dto.input.UserBonusDTO;
-import lotto.dto.input.UserLottoDTO;
-import lotto.dto.input.UserMoneyDTO;
+import lotto.dto.UserBonusDTO;
+import lotto.dto.UserLottoDTO;
+import lotto.dto.UserMoneyDTO;
 import lotto.io.Reader;
 import lotto.io.Writer;
 import lotto.repository.LottoRepository;
