+++ src/test/java/lotto/domain/UserMoneyTest.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
-import lotto.dto.input.UserMoneyDTO;
+import lotto.dto.UserMoneyDTO;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
