+++ src/test/java/lotto/dto/UserBonusDTOTest.java
@@ -3,7 +3,6 @@ package lotto.dto;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.dto.input.UserBonusDTO;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
